+++ src/aux-common/bots/BotEvents.ts
@@ -1850,9 +1850,24 @@ export interface SerialConnectAction extends AsyncAction {
     type: 'serial_connect';
 
     /**
+     * A friendly device name. Example: Brush01
+     */
+    name: string;
+
+    /**
      * The device path. Example: /dev/rfcomm0
      */
-    path: string;
+    device: string;
+
+    /**
+     * The device MAC address. Example: AA:BB:CC:DD:EE
+     */
+    mac: string;
+
+    /**
+     * The device channel. Example: 1
+     */
+    channel: number;
 
     /**
      * {boolean} [autoOpen=true] Automatically opens the port on `nextTick`.
@@ -1888,22 +1903,43 @@ export interface SerialConnectAction extends AsyncAction {
     options?: object;
 }
 /**
- * Parses and returns the serial stream to the event tag 'onStreamData'.
+ * Parses and returns the serial stream to the event tag 'onSerialData'.
  */
 export interface SerialStreamAction extends AsyncAction {
     type: 'serial_stream';
+
+    /**
+     * The id of the bot you want data streamed to. The bot needs the 'onSerialData' tag.
+     */
+    bot: string;
+
+    /**
+     * A friendly device name. Example: Brush01
+     */
+     name: string;
+
 }
 /**
  * Opens the serial connection if you set the option in serialConnect to {autoOpen: false}
  */
 export interface SerialOpenAction extends AsyncAction {
     type: 'serial_open';
+    /**
+     * A friendly device name. Example: Brush01
+     */
+    name: string;
+
 }
 /**
  * Updates the SerialPort object with a new baudRate.
  */
 export interface SerialUpdateAction extends AsyncAction {
     type: 'serial_update';
+    /**
+     * A friendly device name. Example: Brush01
+     */
+    name: string;
+
 
     /**
      * {number=} [baudRate=9600] The baud rate of the port to be opened. This should match one of the commonly available baud rates, such as 110, 300, 1200, 2400, 4800, 9600, 14400, 19200, 38400, 57600, or 115200. Custom rates are supported best effort per platform. The device connected to the serial port is not guaranteed to support the requested baud rate, even if the port itself supports that baud rate.
@@ -1920,6 +1956,11 @@ export interface SerialUpdateAction extends AsyncAction {
  */
 export interface SerialWriteAction extends AsyncAction {
     type: 'serial_write';
+    /**
+     * A friendly device name. Example: Brush01
+     */
+    name: string;
+
 
     /**
      * The data/command to send.
@@ -1941,6 +1982,11 @@ export interface SerialWriteAction extends AsyncAction {
  */
 export interface SerialReadAction extends AsyncAction {
     type: 'serial_read';
+    /**
+     * A friendly device name. Example: Brush01
+     */
+    name: string;
+
 
     /**
      * Specify how many bytes of data to return, if available
@@ -1952,6 +1998,16 @@ export interface SerialReadAction extends AsyncAction {
  */
 export interface SerialCloseAction extends AsyncAction {
     type: 'serial_close';
+    /**
+     * A friendly device name. Example: Brush01
+     */
+    name: string;
+
+
+    /**
+     * The device path. Example: /dev/rfcomm0
+     */
+     device: string;
 
     /**
      *
@@ -1963,24 +2019,44 @@ export interface SerialCloseAction extends AsyncAction {
  */
 export interface SerialFlushAction extends AsyncAction {
     type: 'serial_flush';
+    /**
+     * A friendly device name. Example: Brush01
+     */
+    name: string;
+
 }
 /**
  * Waits until all output data is transmitted to the serial port. After any pending write has completed, it calls `tcdrain()` or `FlushFileBuffers()` to ensure it has been written to the device.
  */
 export interface SerialDrainAction extends AsyncAction {
     type: 'serial_drain';
+    /**
+     * A friendly device name. Example: Brush01
+     */
+    name: string;
+
 }
 /**
  * Causes a stream in flowing mode to stop emitting 'data' events, switching out of flowing mode. Any data that becomes available remains in the internal buffer.
  */
 export interface SerialPauseAction extends AsyncAction {
     type: 'serial_pause';
+    /**
+     * A friendly device name. Example: Brush01
+     */
+    name: string;
+
 }
 /**
  * Causes an explicitly paused, Readable stream to resume emitting 'data' events, switching the stream into flowing mode.
  */
 export interface SerialResumeAction extends AsyncAction {
     type: 'serial_resume';
+    /**
+     * A friendly device name. Example: Brush01
+     */
+    name: string;
+
 }
 /**
  * Defines an event that sets some text on the user's clipboard.
@@ -4303,6 +4379,7 @@ export function rpioSPIEndPin(
 
 /**
  * Establish the connection to the bluetooth serial device
+ * @param name A friendly device name. Example: Brush01
  * @param path The device path. Example: /dev/rfcomm0
  * @param options
  * {boolean} [autoOpen=true] Automatically opens the port on `nextTick`.
@@ -4337,13 +4414,19 @@ export function rpioSPIEndPin(
  * @param taskId The ID of the async task.
  */
 export function serialConnectPin(
-    path: string,
+    name: string,
+    device: string,
+    mac: string,
+    channel: number,
     options?: object,
     taskId?: string | number,
     playerId?: string
 ): SerialConnectAction {
     return {
-        path,
+        name,
+        device,
+        mac,
+        channel,
         options,
         type: 'serial_connect',
         taskId,
@@ -4352,14 +4435,20 @@ export function serialConnectPin(
 }
 
 /**
- * Parses and returns the serial stream to the event tag 'onStreamData'.
+ * Parses and returns the serial stream to the event tag 'onSerialData'.
+ * @param bot The id of the bot you want data streamed to. The bot needs the 'onSerialData' tag.
+ * @param name A friendly device name. Example: Brush01
  * @param taskId The ID of the async task.
  */
 export function serialStreamPin(
+    bot: string,
+    name: string,
     taskId?: string | number,
     playerId?: string
 ): SerialStreamAction {
     return {
+        bot,
+        name,
         type: 'serial_stream',
         taskId,
         playerId,
@@ -4368,13 +4457,16 @@ export function serialStreamPin(
 
 /**
  * Opens the serial connection if you set the option in serialConnect to {autoOpen: false}
+ * @param name A friendly device name. Example: Brush01
  * @param taskId The ID of the async task.
  */
 export function serialOpenPin(
+    name: string,
     taskId?: string | number,
     playerId?: string
 ): SerialOpenAction {
     return {
+        name,
         type: 'serial_open',
         taskId,
         playerId,
@@ -4383,17 +4475,20 @@ export function serialOpenPin(
 
 /**
  * Updates the SerialPort object with a new baudRate.
+ * @param name A friendly device name. Example: Brush01
  * @param options {number=} [baudRate=9600] The baud rate of the port to be opened. This should match one of the commonly available baud rates, such as 110, 300, 1200, 2400, 4800, 9600, 14400, 19200, 38400, 57600, or 115200. Custom rates are supported best effort per platform. The device connected to the serial port is not guaranteed to support the requested baud rate, even if the port itself supports that baud rate.
  * @param cb
  * @param taskId The ID of the async task.
  */
 export function serialUpdatePin(
+    name: string,
     options: object,
     cb?: any,
     taskId?: string | number,
     playerId?: string
 ): SerialUpdateAction {
     return {
+        name,
         options,
         cb,
         type: 'serial_update',
@@ -4404,12 +4499,14 @@ export function serialUpdatePin(
 
 /**
  * Writes the provided data/command to the device
+ * @param name A friendly device name. Example: Brush01
  * @param data The data/command to send
  * @param encoding The encoding, if chunk is a string. Defaults to 'utf8'. Also accepts 'utf16le', 'latin1', 'ascii', 'base64', 'binary', 'ucs2', and 'hex'
  * @param cb
  * @param taskId The ID of the async task.
  */
 export function serialWritePin(
+    name: string,
     data: string | number[],
     encoding?: string,
     cb?: any,
@@ -4417,6 +4514,7 @@ export function serialWritePin(
     playerId?: string
 ): SerialWriteAction {
     return {
+        name,
         data,
         encoding,
         cb,
@@ -4428,15 +4526,18 @@ export function serialWritePin(
 
 /**
  * Request a number of bytes from the SerialPort.
+ * @param name A friendly device name. Example: Brush01
  * @param size Specify how many bytes of data to return, if available.
  * @param taskId The ID of the async task.
  */
 export function serialReadPin(
+    name: string,
     size?: number,
     taskId?: string | number,
     playerId?: string
 ): SerialReadAction {
     return {
+        name,
         size,
         type: 'serial_read',
         taskId,
@@ -4446,15 +4547,21 @@ export function serialReadPin(
 
 /**
  * Closes an open connection.
+ * @param name A friendly device name. Example: Brush01
  * @param cb
+ * @param device The device path. Example: /dev/rfcomm0
  * @param taskId The ID of the async task.
  */
 export function serialClosePin(
+    name: string,
+    device: string,
     cb?: any,
     taskId?: string | number,
     playerId?: string
 ): SerialCloseAction {
     return {
+        name,
+        device,
         cb,
         type: 'serial_close',
         taskId,
@@ -4464,13 +4571,16 @@ export function serialClosePin(
 
 /**
  * Flush discards data that has been received but not read, or written but not transmitted by the operating system.
+ * @param name A friendly device name. Example: Brush01
  * @param taskId The ID of the async task.
  */
 export function serialFlushPin(
+    name: string,
     taskId?: string | number,
     playerId?: string
 ): SerialFlushAction {
     return {
+        name,
         type: 'serial_flush',
         taskId,
         playerId,
@@ -4479,13 +4589,16 @@ export function serialFlushPin(
 
 /**
  * Waits until all output data is transmitted to the serial port. After any pending write has completed, it calls `tcdrain()` or `FlushFileBuffers()` to ensure it has been written to the device.
+ * @param name A friendly device name. Example: Brush01
  * @param taskId The ID of the async task.
  */
 export function serialDrainPin(
+    name: string,
     taskId?: string | number,
     playerId?: string
 ): SerialDrainAction {
     return {
+        name,
         type: 'serial_drain',
         taskId,
         playerId,
@@ -4494,13 +4607,16 @@ export function serialDrainPin(
 
 /**
  * Causes a stream in flowing mode to stop emitting 'data' events, switching out of flowing mode. Any data that becomes available remains in the internal buffer.
+ * @param name A friendly device name. Example: Brush01
  * @param taskId The ID of the async task.
  */
 export function serialPausePin(
+    name: string,
     taskId?: string | number,
     playerId?: string
 ): SerialPauseAction {
     return {
+        name,
         type: 'serial_pause',
         taskId,
         playerId,
@@ -4509,13 +4625,16 @@ export function serialPausePin(
 
 /**
  * Causes an explicitly paused, Readable stream to resume emitting 'data' events, switching the stream into flowing mode.
+ * @param name A friendly device name. Example: Brush01
  * @param taskId The ID of the async task.
  */
 export function serialResumePin(
+    name: string,
     taskId?: string | number,
     playerId?: string
 ): SerialResumeAction {
     return {
+        name,
         type: 'serial_resume',
         taskId,
         playerId,