+++ src/aux-common/runtime/AuxLibrary.spec.ts
@@ -4035,11 +4035,14 @@ describe('AuxLibrary', () => {
             it('should send a SerialConnectAction in a RemoteAction', () => {
                 uuidMock.mockReturnValueOnce('task1');
                 const action: any = library.api.server.serialConnect(
+                    'Brush01',
                     '/dev/ttyS0',
+                    'AA:BB:CC:DD:EE',
+                    1,
                     { baudRate: 9600 }
                 );
                 const expected = remote(
-                    serialConnectPin('/dev/ttyS0', { baudRate: 9600 }),
+                    serialConnectPin('Brush01','/dev/ttyS0','AA:BB:CC:DD:EE', 1, { baudRate: 9600 }),
                     undefined,
                     undefined,
                     'task1'
@@ -4050,9 +4053,7 @@ describe('AuxLibrary', () => {
 
             it('should create tasks that can be resolved from a remote', () => {
                 uuidMock.mockReturnValueOnce('uuid');
-                library.api.server.serialConnect('/dev/ttyS0', {
-                    baudRate: 9600,
-                });
+                library.api.server.serialConnect('Brush01','/dev/ttyS0','AA:BB:CC:DD:EE', 1, { baudRate: 9600 });
 
                 const task = context.tasks.get('uuid');
                 expect(task.allowRemoteResolution).toBe(true);
@@ -4062,9 +4063,9 @@ describe('AuxLibrary', () => {
         describe('server.serialStream()', () => {
             it('should send a SerialStreamAction in a RemoteAction', () => {
                 uuidMock.mockReturnValueOnce('task1');
-                const action: any = library.api.server.serialStream();
+                const action: any = library.api.server.serialStream('1a2b3', 'Brush01');
                 const expected = remote(
-                    serialStreamPin(),
+                    serialStreamPin('1a2b3', 'Brush01'),
                     undefined,
                     undefined,
                     'task1'
@@ -4075,7 +4076,7 @@ describe('AuxLibrary', () => {
 
             it('should create tasks that can be resolved from a remote', () => {
                 uuidMock.mockReturnValueOnce('uuid');
-                library.api.server.serialStream();
+                library.api.server.serialStream('1a2b3', 'Brush01');
 
                 const task = context.tasks.get('uuid');
                 expect(task.allowRemoteResolution).toBe(true);
@@ -4085,9 +4086,9 @@ describe('AuxLibrary', () => {
         describe('server.serialOpen()', () => {
             it('should send a SerialOpenAction in a RemoteAction', () => {
                 uuidMock.mockReturnValueOnce('task1');
-                const action: any = library.api.server.serialOpen();
+                const action: any = library.api.server.serialOpen('Brush01');
                 const expected = remote(
-                    serialOpenPin(),
+                    serialOpenPin('Brush01'),
                     undefined,
                     undefined,
                     'task1'
@@ -4098,7 +4099,7 @@ describe('AuxLibrary', () => {
 
             it('should create tasks that can be resolved from a remote', () => {
                 uuidMock.mockReturnValueOnce('uuid');
-                library.api.server.serialOpen();
+                library.api.server.serialOpen('Brush01');
 
                 const task = context.tasks.get('uuid');
                 expect(task.allowRemoteResolution).toBe(true);
@@ -4108,11 +4109,11 @@ describe('AuxLibrary', () => {
         describe('server.serialUpdate()', () => {
             it('should send a SerialUpdateAction in a RemoteAction', () => {
                 uuidMock.mockReturnValueOnce('task1');
-                const action: any = library.api.server.serialUpdate({
+                const action: any = library.api.server.serialUpdate('Brush01', {
                     baudRate: 9600,
                 });
                 const expected = remote(
-                    serialUpdatePin({ baudRate: 9600 }),
+                    serialUpdatePin('Brush01', { baudRate: 9600 }),
                     undefined,
                     undefined,
                     'task1'
@@ -4123,7 +4124,7 @@ describe('AuxLibrary', () => {
 
             it('should create tasks that can be resolved from a remote', () => {
                 uuidMock.mockReturnValueOnce('uuid');
-                library.api.server.serialUpdate({ baudRate: 9600 });
+                library.api.server.serialUpdate('Brush01', { baudRate: 9600 });
 
                 const task = context.tasks.get('uuid');
                 expect(task.allowRemoteResolution).toBe(true);
@@ -4134,11 +4135,12 @@ describe('AuxLibrary', () => {
             it('should send a SerialWriteAction in a RemoteAction', () => {
                 uuidMock.mockReturnValueOnce('task1');
                 const action: any = library.api.server.serialWrite(
+                    'Brush01',
                     'Hello World!',
                     'utf8'
                 );
                 const expected = remote(
-                    serialWritePin('Hello World!', 'utf8'),
+                    serialWritePin('Brush01', 'Hello World!', 'utf8'),
                     undefined,
                     undefined,
                     'task1'
@@ -4149,7 +4151,7 @@ describe('AuxLibrary', () => {
 
             it('should create tasks that can be resolved from a remote', () => {
                 uuidMock.mockReturnValueOnce('uuid');
-                library.api.server.serialWrite('Hello World!', 'utf8');
+                library.api.server.serialWrite('Brush01', 'Hello World!', 'utf8');
 
                 const task = context.tasks.get('uuid');
                 expect(task.allowRemoteResolution).toBe(true);
@@ -4159,9 +4161,9 @@ describe('AuxLibrary', () => {
         describe('server.serialRead()', () => {
             it('should send a SerialReadAction in a RemoteAction', () => {
                 uuidMock.mockReturnValueOnce('task1');
-                const action: any = library.api.server.serialRead();
+                const action: any = library.api.server.serialRead('Brush01');
                 const expected = remote(
-                    serialReadPin(),
+                    serialReadPin('Brush01'),
                     undefined,
                     undefined,
                     'task1'
@@ -4172,7 +4174,7 @@ describe('AuxLibrary', () => {
 
             it('should create tasks that can be resolved from a remote', () => {
                 uuidMock.mockReturnValueOnce('uuid');
-                library.api.server.serialRead();
+                library.api.server.serialRead('Brush01');
 
                 const task = context.tasks.get('uuid');
                 expect(task.allowRemoteResolution).toBe(true);
@@ -4182,9 +4184,9 @@ describe('AuxLibrary', () => {
         describe('server.serialClose()', () => {
             it('should send a SerialCloseAction in a RemoteAction', () => {
                 uuidMock.mockReturnValueOnce('task1');
-                const action: any = library.api.server.serialClose();
+                const action: any = library.api.server.serialClose('Brush01', "/dev/rfcomm0");
                 const expected = remote(
-                    serialClosePin(),
+                    serialClosePin('Brush01', "/dev/rfcomm0"),
                     undefined,
                     undefined,
                     'task1'
@@ -4195,7 +4197,7 @@ describe('AuxLibrary', () => {
 
             it('should create tasks that can be resolved from a remote', () => {
                 uuidMock.mockReturnValueOnce('uuid');
-                library.api.server.serialClose();
+                library.api.server.serialClose('Brush01', "/dev/rfcomm0");
 
                 const task = context.tasks.get('uuid');
                 expect(task.allowRemoteResolution).toBe(true);
@@ -4205,9 +4207,9 @@ describe('AuxLibrary', () => {
         describe('server.serialFlush()', () => {
             it('should send a SerialFlushAction in a RemoteAction', () => {
                 uuidMock.mockReturnValueOnce('task1');
-                const action: any = library.api.server.serialFlush();
+                const action: any = library.api.server.serialFlush('Brush01');
                 const expected = remote(
-                    serialFlushPin(),
+                    serialFlushPin('Brush01'),
                     undefined,
                     undefined,
                     'task1'
@@ -4218,7 +4220,7 @@ describe('AuxLibrary', () => {
 
             it('should create tasks that can be resolved from a remote', () => {
                 uuidMock.mockReturnValueOnce('uuid');
-                library.api.server.serialFlush();
+                library.api.server.serialFlush('Brush01');
 
                 const task = context.tasks.get('uuid');
                 expect(task.allowRemoteResolution).toBe(true);
@@ -4228,9 +4230,9 @@ describe('AuxLibrary', () => {
         describe('server.serialDrain()', () => {
             it('should send a SerialDrainAction in a RemoteAction', () => {
                 uuidMock.mockReturnValueOnce('task1');
-                const action: any = library.api.server.serialDrain();
+                const action: any = library.api.server.serialDrain('Brush01');
                 const expected = remote(
-                    serialDrainPin(),
+                    serialDrainPin('Brush01'),
                     undefined,
                     undefined,
                     'task1'
@@ -4241,7 +4243,7 @@ describe('AuxLibrary', () => {
 
             it('should create tasks that can be resolved from a remote', () => {
                 uuidMock.mockReturnValueOnce('uuid');
-                library.api.server.serialDrain();
+                library.api.server.serialDrain('Brush01');
 
                 const task = context.tasks.get('uuid');
                 expect(task.allowRemoteResolution).toBe(true);
@@ -4251,9 +4253,9 @@ describe('AuxLibrary', () => {
         describe('server.serialPause()', () => {
             it('should send a SerialPauseAction in a RemoteAction', () => {
                 uuidMock.mockReturnValueOnce('task1');
-                const action: any = library.api.server.serialPause();
+                const action: any = library.api.server.serialPause('Brush01');
                 const expected = remote(
-                    serialPausePin(),
+                    serialPausePin('Brush01'),
                     undefined,
                     undefined,
                     'task1'
@@ -4264,7 +4266,7 @@ describe('AuxLibrary', () => {
 
             it('should create tasks that can be resolved from a remote', () => {
                 uuidMock.mockReturnValueOnce('uuid');
-                library.api.server.serialPause();
+                library.api.server.serialPause('Brush01');
 
                 const task = context.tasks.get('uuid');
                 expect(task.allowRemoteResolution).toBe(true);
@@ -4274,9 +4276,9 @@ describe('AuxLibrary', () => {
         describe('server.serialResume()', () => {
             it('should send a SerialResumeAction in a RemoteAction', () => {
                 uuidMock.mockReturnValueOnce('task1');
-                const action: any = library.api.server.serialResume();
+                const action: any = library.api.server.serialResume('Brush01');
                 const expected = remote(
-                    serialResumePin(),
+                    serialResumePin('Brush01'),
                     undefined,
                     undefined,
                     'task1'
@@ -4287,7 +4289,7 @@ describe('AuxLibrary', () => {
 
             it('should create tasks that can be resolved from a remote', () => {
                 uuidMock.mockReturnValueOnce('uuid');
-                library.api.server.serialResume();
+                library.api.server.serialResume('Brush01');
 
                 const task = context.tasks.get('uuid');
                 expect(task.allowRemoteResolution).toBe(true);